Rename style file, event listener
