+++ src/styles/item-image-styles.ts
